+++ event-sourcing/event-publisher/event-publisher-timer/src/main/java/uk/gov/justice/services/eventsourcing/publishedevent/prepublish/PrePublisherTimerBean.java
@@ -28,6 +28,9 @@ public class PrePublisherTimerBean {
     @Inject
     private AsynchronousPrePublisher asynchronousPrePublisher;
 
+    @Inject
+    private PrePublishNotifier prePublishNotifier;
+
     @PostConstruct
     public void startTimerService() {
 
@@ -42,7 +45,12 @@ public class PrePublisherTimerBean {
     public void performPrePublish() {
 
         if (! prePublisherTimerConfig.isDisabled()) {
-            asynchronousPrePublisher.performPrePublish();
+            if (prePublisherTimerConfig.shouldWorkerNotified())
+                prePublishNotifier.wakeUp(true);
+            else {
+                asynchronousPrePublisher.performPrePublish();
+            }
+
         }
     }
 }